Refactor measurement observability package
